requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
et-xmlfile==1.1.0
h11==0.14.0
idna==3.4
jsonschema==4.17.3
lxml @ file:///C:/GHPY/lxml-4.9.0-cp311-cp311-win_amd64.whl
numpy==1.24.1
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
pdfminer.six==20221105
Pillow==9.3.0
pycparser==2.21
pypiwin32==223
pyproject-toml==0.0.10
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.7
pywin32==305
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
